Now using .bashrc instead of .bash_profile, minor updates to scripts

+++ add.sh
@@ -1,12 +1,17 @@
 ## declare an array variable
-declare -a files=(".bash_aliases" ".bash_profile" ".git-prompt.sh" ".gitconfig")
+declare -a files=(".bashrc" ".bash_aliases" ".bash_profile" ".git-prompt.sh" ".gitconfig")
 
 ## now loop through the above array
 for file in "${files[@]}"
 do
 	if [[ -e "$HOME/$file" ]]
 	then
-		mv "$HOME/$file" "$HOME/$file.old"
+		if [[ -e "$HOME/$file.old"]]
+		then
+			echo "$HOME/$file.old already exists, won't replace it"
+		else
+			mv "$HOME/$file" "$HOME/$file.old"
+		fi
 	fi
 	cp "$file" "$HOME/$file"
 	# or do whatever with individual element of the array
